requirements.txt
langchain-ollama
langchain
ipykernel
python-dotenv
langchain_community
langchain-community
langchain_core
streamlit
langchain_groq
faiss-cpu
langchain_huggingface
pypdf
beautifulsoup4  
arxiv 
pymupdf
wikipedia
langchain-openai
chromadb
sentence-transformers  
duckdb
pandas
duckduckgo-search
mysql-connector-python
SQLAlchemy
validators==0.28.1  #To validate URLs
youtube-transcript-api  #To read transcript of the YouTube video
pytube
numexpr
unstructured
huggingface_hub
mistral_inference
PyPDF2
astrapy
cassio
datasets
tiktoken
gradio
crewai
crewai_tools